(Breaking) Make Lua.Table.as_string more powerful
Changes the API of `Lua.Table.as_string/2` so that it can take a
`:formatter` option, which can format any value prior to serialization.
The function is a 2-arity function that takes the key and value, and can
transform the value prior to serialization.

Also allows `as_string/2` to take lists